requirements.txt
adbc_driver_manager==1.2.0
adbc_driver_postgresql==1.2.0
adbc_driver_sqlite==1.2.0
anywidget==0.9.13
AppKit==0.2.8
atheris==2.3.0
beautifulsoup4==4.12.3
bokeh==3.6.0
boto3==1.35.34
botocore==1.35.34
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
checks==0.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==43.0.1
cudf==0.6.1.post1
cycler==0.12.1
Cython==3.0.11
dask==2024.9.1
dask_expr==1.1.15
defusedxml==0.7.1
docutils==0.21.2
duckdb==1.1.1
fastparquet==2024.5.0
filelock==3.16.1
Foundation==0.1.0a0.dev1
fqdn==1.5.1
fsspec==2024.9.0
gitdb_speedups==0.1.0
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.112.4
ibis==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jnius==1.1.0
JPype1==1.5.0
jsonpointer==3.0.0
keyring==25.4.1
lxml==5.3.0
matplotlib==3.9.2
memory_profiler==0.61.0
modin==0.32.0
mtrand==0.1
mypy==1.11.2
numexpr==2.10.1
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
pickle5==0.0.12
pkgutil_resolve_name==1.3.10
plotly==5.24.1
polars==1.9.0
psutil==6.0.0
py==1.11.0
pycares==4.4.0
pycurl==7.45.3
PyInstaller==6.10.0
pyodide==0.0.2
pyOpenSSL==24.2.1
pyperf==2.8.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.7.1
PySide6==6.7.3
pytest==8.3.3
python_calamine==0.2.3
pyxlsb==1.0.10
PyYAML==6.0.2
PyYAML==6.0.2
qtpy==2.4.1
redis==5.1.1
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
s3fs==2024.9.0
scikit_learn==1.5.2
scipy==1.14.1
scipy_doctest==1.4
selenium==4.25.0
sets==0.3.2
setuptools==75.1.0
setuptools_scm==8.1.0
simplejson==3.19.3
sniffio==1.3.1
snowflake==0.13.0
Sphinx==8.0.2
SQLAlchemy==2.0.35
st_annotated_text==4.0.1
streamlit_ace==0.1.1
sympy==1.13.3
tables==3.10.1
testbench==0.1.2
thread==2.0.5
threadpoolctl==3.5.0
traitlets==5.14.3
trio==0.26.2
uri_template==1.3.0
urllib3_secure_extra==0.1.0
vegafusion==1.6.9
watchdog==5.0.3
webcolors==24.8.0
xarray==2024.9.0
xlrd==2.0.1
xlsxwriter==3.2.0
xmlrpclib==1.0.1
zstandard==0.23.0
